# openapi/openapi.yaml
openapi: "3.1.0"
info:
  title: UpdateDeck Internal HTTP API
  version: "0.1.0"
  description: |
    Cookie-session job and ops endpoints. The product UI primarily uses
    Next.js Server Actions (`src/lib/actions/`). These HTTP routes mirror
    enqueue operations for scripts, webhooks consumers, and future clients.

    Authentication for deck routes: Supabase SSR session cookie
    (same-origin browser or cookie jar). No Bearer API keys yet.
servers:
  - url: http://localhost:3000
    description: Local
  - url: "{origin}"
    variables:
      origin:
        default: https://app.example.com
...
paths:
...
  /api/webhooks/stripe:
    post:
      tags: [Webhooks]
      operationId: stripeWebhook
      summary: Stripe billing events
      security: []
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Event received
        "400":
          description: Invalid signature or payload
...
components:
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase SSR auth cookies (browser session)
security:
  - SessionCookie: []